Moved all force calculations into forces.c
