+++ MANIFEST.in
@@ -1,10 +1,9 @@
 include src/assist.c
-include src/constants.c
-include src/get_mass.c
+include src/forces.c
 include src/planets.c
 include src/spk.c
-include src/spkhead.c
 include src/assist.h
+include src/forces.h
 include src/const.h
 include src/planets.h
 include src/spk.h
